LRAC-5080 Another race condition part 2. Our logic in onAfterUpdate expected the actual AnalyticsConfiguration to be updated in time. Avoid these timing dependencies by consolidating the logic into the tracker.
